+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 0.1.9
 
   * Fix fetching of members for a waiting list
+  * Fix 'undefuned method' when creating an event without a start or finish datetime
 
 ## Version 0.1.8